create link to exiftool binary

+++ extragear/digikam/digikam.py
@@ -461,6 +461,12 @@ class Package(CMakePackageBase):
                 print("Could not rename ExifTool directory")
                 return False
 
+            if not utils.createSymlink(os.path.join("./", "Image-ExifTool", "exiftool"),
+                                       os.path.join(self.archiveDir(), "bin", "exiftool"),
+                                       targetIsDirectory=False):
+                print("Could not create symlink to ExifTool binary")
+                return False
+
             if not utils.deleteFile(os.path.join(binPath, "Image-ExifTool.tar.gz")):
                 print("Could not remove ExifTool archive")
                 return False
